pyproject.toml
[tool.poetry]
name = "cf2tf"
version = "0.1.0"
description = "Convert Cloudformation Templates to Terraform."
authors = ["Levi Blaney <shadycuz@gmail.com>"]

[tool.poetry.scripts]
cf2tf = "cf2tf.app:cli"

[tool.poetry.dependencies]
python = "^3.9"
cfn-flip = "^1.3.0"
PyYAML = "^6.0"
click = "^8.1.2"
GitPython = "^3.1.27"
thefuzz = {extras = ["speedup"], version = "^0.19.0"}
click-log = "^0.4.0"
requests = "^2.27.1"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
flake8-black = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
